Add files via upload
Selore Nigeria is a popular electronics retail chain in
Nigeria that specializes in the sale of mobile phones,
tablets, laptops, and other related accessories. The
company was founded in 2022 and has since grown to
become one of the largest retailers of mobile phones and
related products in the country.
Selore Nigeria offers a wide range of mobile phones from
top brands such as Samsung, Apple, Nokia, and Huawei,

among others. The company is also known for its after-
sales support, including repairs and maintenance services,

as well as trade-in options for old devices. With its
numerous outlets nationwide, Selore Nigeria is a go-to
destination for many Nigerians seeking to purchase mobile
phones and related accessories.

A retail company wants to understand the spending
behaviour of its customers to target them with personalized
marketing campaigns. They have collected data on
customer demographics (age, gender) and spending habits
(annual income, spending score) and want to segment their
customers based on their spending behaviour. The
objective of this project is to build a customer segmentation
model to help the retail company in developing targeted
marketing strategies.
Data Collection:
The retail company has collected data on 200 customers
containing
• CustomerID,
• Gender,
• Age,
• Annual_Income_(k$),
• Spending_Score.

The goal of this project is to segment
customers based on their purchasing
behaviour and demographic information.
We will use unsupervised learning
techniques to identify groups of customers
with similar behaviour and
characteristics. This information can be
used to develop targeted marketing
campaigns, personalize product
recommendations, and optimize pricing
strategies.